+++ src/wildtornado/scocalc/strategies/CommentPercentageCalculator.java
@@ -2,12 +2,12 @@ package wildtornado.scocalc.strategies;
 
 public class CommentPercentageCalculator extends BaseCalculator {
 
-    public CommentPercentageCalculator(int val) {
+    public CommentPercentageCalculator(double val) {
         this.val = val;
     }
 
     public void generate() {
-        this.result = val;
+        this.result = (int) val;
     }
 
 }
